Extract generic type closing to a separate class
It's still not extensible somehow - it was extracted to a separate class
for easier maintenance. It's a whole separate concern.
